Add "zset-low-rank" and "zset-high-rank" maxmemory eviction policies
Instead of evicting entire keys based on LRU, TTL, or simply selecting
random keys, these new policies remove the lowest or highest ranked
*member* from sorted sets to free memory.

The idea is that we can then continue to add members to existing keys and
ensure the Redis instance is using an upper-bounded amount of memory. Using
any of the existing eviction policies would always result in the entire key
(and thus all the members) being evicted, which would be premature and sub-
optimal in many use-cases.

We log the number of member evictions to a new "evicted_zsetmembers" entry
in the INFO command. If we end up removing the last member from a set, we
remove the key entirely. This is reflected in the existing evicted_keys
INFO entry.

Note that we sample maxmemory-samples keys from the entire database - if
this selects keys that are not sorted sets (eg. hashes, strings, etc.) then
they still "count" as a sample. Therefore, if you have a mix of data types
in your db, you may need to increase maxmemory-samples or move the sorted
sets to their own database.

Testsuite additions are included in this patch.

To demo:

  - ZADD key 1 one 2 two 3 three 4 four

  - CONFIG SET maxmemory-policy zset-low-rank

  - CONFIG SET maxmemory 1

  - ZRANGE key 0 -1 WITHSCORES
    1) "two"
    2) "2"
    3) "three"
    4) "3"
    5) "four"
    6) "4"

  - ZRANGE key 0 -1 WITHSCORES
    1) "three"
    2) "3"
    3) "four"
    4) "4"

  - ZRANGE key 0 -1 WITHSCORES
    1) "four"
    2) "4"

  - ZRANGE key 0 -1 WITHSCORES
    (empty list or set)

  - INFO | grep evicted
    evicted_keys:1
    evicted_zsetmembers:4

zset-high-rank would look identical except that the set members would be
removed in reverse order.

Signed-off-by: Chris Lamb <chris@chris-lamb.co.uk>